+++ lib/ExtensionManager.d.ts
@@ -1,14 +1,38 @@
-import InclusionObject from './objects/InclusionObject';
+import HtmlConverter from './HtmlConverter';
+import MarkdownConverter from './MarkdownConverter';
+import ExtensionObject from './objects/ExtensionObject';
 /**
 * Allows to parse HTML code for usage of elearn.js extensions.
 * Allows to get path to extension necessary files.
 * Allows to get include strings for extension necessary files.
 */
 declare class ExtensionManager {
+    static htmlConverter: HtmlConverter;
     static scanForQuiz(html: string): boolean;
     static scanForVideo(html: string): boolean;
     static scanForClickImage(html: string): boolean;
     static scanForTimeSlider(html: string): boolean;
+    /**
+     * Scans the html code for elearn.js extensions.
+     *
+     * @param html string: the html code to be scanned
+     *
+     * @return ExtensionObject: including which extensions where found
+     * and explicitly which were not found (true/false)
+     */
+    static scanHtmlForAll(html: string): ExtensionObject;
+    /**
+     * Scans the markdown code for elearn.js extensions.
+     *
+     * @param markdown string: the original markdown code to be scanned
+     * @param markdownConverter (optional) HtmlConverter or PdfConverter:
+     *  the converter to use for markdown to HTML conversion. If not given,
+     *  a default HtmlConverter will be used.
+     *
+     * @return Promise<ExtensionObject>: including which extensions where found
+     * and explicitly which were not found (true/false)
+     */
+    static scanMarkdownForAll(markdown: string, markdownConverter?: MarkdownConverter): Promise<ExtensionObject>;
     static getQuizAssetDir(): string;
     static getElearnVideoAssetDir(): string;
     static getClickImageAssetDir(): string;
@@ -27,9 +51,9 @@ declare class ExtensionManager {
     /**
     * Writes the elearn.js assets to the given path.
     * @param dirPath: string - the path to write the `assets` folder to.
-    * @param {InclusionObject} opts optional options
+    * @param {ExtensionObject} opts optional options
     */
-    static exportAssets(dirPath: string, opts: InclusionObject): Promise<{}>;
+    static exportAssets(dirPath: string, opts: ExtensionObject): Promise<{}>;
     /**
     * Copies/writes a list of folders by their absolute paths to the outPath
     */