+++ test/util/PromiseCounter.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const PromiseCounter = require("../../src/util/PromiseCounter.js");
+const PromiseCounter = require("../../lib/util/PromiseCounter");
 
 /**
 * Creates `count` Promises.
@@ -18,10 +18,10 @@ var createPromises = function(count, resolving, timeoutMin, timeoutMax) {
 
     var promises = [];
 
-    for(var i=0; i<count; i++) {
+    for(var i = 0; i < count; i++) {
         var promise = new Promise((resolve, reject) => {
-            if(resolving) setTimeout(resolve, timeoutBase + (i*timeoutStep));
-            else setTimeout(reject, timeoutBase + (i*timeoutStep));
+            if(resolving) setTimeout(resolve, timeoutBase + (i * timeoutStep));
+            else setTimeout(reject, timeoutBase + (i * timeoutStep));
         });
         promises.push(promise);
     }
@@ -30,18 +30,19 @@ var createPromises = function(count, resolving, timeoutMin, timeoutMax) {
 }
 
 module.exports.resolvedPromises = (test) => {
+    console.log(PromiseCounter);
     var promises = createPromises(20, true, 0, 5000);
 
     test.expect(1);
 
     var counter = new PromiseCounter(promises);
     counter.then(() => {
-            test.ok(true);
-            test.done();
-        }, (err) => {
-            throw "Promises not correctly resolved.";
-            test.done();
-        });
+        test.ok(true);
+        test.done();
+    }, (err) => {
+        throw "Promises not correctly resolved.";
+        test.done();
+    });
 }
 
 module.exports.rejectedPromises = (test) => {
@@ -51,12 +52,12 @@ module.exports.rejectedPromises = (test) => {
 
     var counter = new PromiseCounter(promises, 10000);
     counter.then(() => {
-            throw "Promises not correctly rejected.";
-            test.done();
-        }, (err) => {
-            test.ok(true);
-            test.done();
-        });
+        throw "Promises not correctly rejected.";
+        test.done();
+    }, (err) => {
+        test.ok(true);
+        test.done();
+    });
 }
 
 module.exports.oneTimeouted = (test) => {
@@ -71,12 +72,12 @@ module.exports.oneTimeouted = (test) => {
 
     var counter = new PromiseCounter(promises, timeout);
     counter.then(() => {
-            throw "Promises not correctly rejected.";
-            test.done();
-        }, (err) => {
-            test.equals(err, `Timeout in PromiseCounter after ${timeout} ms.`);
-            test.done();
-        });
+        throw "Promises not correctly rejected.";
+        test.done();
+    }, (err) => {
+        test.equals(err, `Timeout in PromiseCounter after ${timeout} ms.`);
+        test.done();
+    });
 }
 
 module.exports.allTimeouted = (test) => {
@@ -87,12 +88,12 @@ module.exports.allTimeouted = (test) => {
 
     var counter = new PromiseCounter(promises, timeout);
     counter.then(() => {
-            throw "Promises not correctly rejected.";
-            test.done();
-        }, (err) => {
-            test.equals(err, `Timeout in PromiseCounter after ${timeout} ms.`);
-            test.done();
-        });
+        throw "Promises not correctly rejected.";
+        test.done();
+    }, (err) => {
+        test.equals(err, `Timeout in PromiseCounter after ${timeout} ms.`);
+        test.done();
+    });
 }
 
 module.exports.incorrectInit = (test) => {